feat: extract p2p upload utils
